flask_app/requirements.txt
Flask
boto3
torch
scikit-learn
joblib
python-dotenv

